python-analytics/requirements.txt
# FastAPI e servidor
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0

# Banco de dados
motor==3.3.2
pymongo==4.6.0

# Análise de dados
pandas==2.1.3
numpy==1.26.2
scipy==1.11.4
scikit-learn==1.3.2
statsmodels==0.14.0

# Visualização (para gerar gráficos se necessário)
matplotlib==3.8.2
seaborn==0.13.0
plotly==5.18.0

# APIs externas
httpx==0.25.2
aiohttp==3.9.1

# Autenticação
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6

# Utilitários
python-dotenv==1.0.0
pytz==2023.3

# Machine Learning (opcional, para previsões)
prophet==1.1.5